+++ alikafka-20190916/alibabacloud_alikafka20190916/models.py
@@ -685,6 +685,39 @@ class CreateConsumerGroupResponse(TeaModel):
         return self
 
 
+class CreatePostPayOrderRequestServerlessConfig(TeaModel):
+    def __init__(
+        self,
+        reserved_publish_capacity: int = None,
+        reserved_subscribe_capacity: int = None,
+    ):
+        self.reserved_publish_capacity = reserved_publish_capacity
+        self.reserved_subscribe_capacity = reserved_subscribe_capacity
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.reserved_publish_capacity is not None:
+            result['ReservedPublishCapacity'] = self.reserved_publish_capacity
+        if self.reserved_subscribe_capacity is not None:
+            result['ReservedSubscribeCapacity'] = self.reserved_subscribe_capacity
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ReservedPublishCapacity') is not None:
+            self.reserved_publish_capacity = m.get('ReservedPublishCapacity')
+        if m.get('ReservedSubscribeCapacity') is not None:
+            self.reserved_subscribe_capacity = m.get('ReservedSubscribeCapacity')
+        return self
+
+
 class CreatePostPayOrderRequestTag(TeaModel):
     def __init__(
         self,
@@ -737,9 +770,11 @@ class CreatePostPayOrderRequest(TeaModel):
         eip_max: int = None,
         io_max: int = None,
         io_max_spec: str = None,
+        paid_type: int = None,
         partition_num: int = None,
         region_id: str = None,
         resource_group_id: str = None,
+        serverless_config: CreatePostPayOrderRequestServerlessConfig = None,
         spec_type: str = None,
         tag: List[CreatePostPayOrderRequestTag] = None,
         topic_quota: int = None,
@@ -773,6 +808,7 @@ class CreatePostPayOrderRequest(TeaModel):
         # *   You must specify at least one of the IoMax and IoMaxSpec parameters. If you configure both parameters, the value of the IoMaxSpec parameter takes effect. We recommend that you specify only the IoMaxSpec parameter.
         # *   For more information about the valid values, see [Billing](~~84737~~).
         self.io_max_spec = io_max_spec
+        self.paid_type = paid_type
         # The number of partitions. We recommend that you configure this parameter.
         # 
         # *   You must specify at least one of the PartitionNum and TopicQuota parameters. We recommend that you configure only the PartitionNum parameter.
@@ -785,6 +821,212 @@ class CreatePostPayOrderRequest(TeaModel):
         # 
         # If this parameter is left empty, the default resource group is used. You can view the resource group ID on the Resource Group page in the Resource Management console.
         self.resource_group_id = resource_group_id
+        self.serverless_config = serverless_config
+        # The edition of the instance. Valid values:
+        # 
+        # *   **normal**: Standard Edition (High Write)
+        # *   **professional**: Professional Edition (High Write)
+        # *   **professionalForHighRead**: Professional Edition (High Read)
+        # 
+        # For more information about these instance editions, see [Billing](~~84737~~).
+        self.spec_type = spec_type
+        # The tags.
+        self.tag = tag
+        # The number of topics. We recommend that you do not configure this parameter.
+        # 
+        # *   You must specify at least one of the PartitionNum and TopicQuota parameters. We recommend that you configure only the PartitionNum parameter.
+        # *   If you specify both parameters, the topic-based sales model is used to check whether the PartitionNum value and the TopicQuota value are the same. If they are not the same, a failure response is returned. If they are the same, the order is placed based on the PartitionNum value.
+        # *   The default value of the TopicQuota parameter varies based on the value of the IoMaxSpec parameter. If the number of topics that you consume exceeds the default value, you are charged additional fees.
+        # *   For more information about the valid values, see [Billing](~~84737~~).
+        self.topic_quota = topic_quota
+
+    def validate(self):
+        if self.serverless_config:
+            self.serverless_config.validate()
+        if self.tag:
+            for k in self.tag:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.deploy_type is not None:
+            result['DeployType'] = self.deploy_type
+        if self.disk_size is not None:
+            result['DiskSize'] = self.disk_size
+        if self.disk_type is not None:
+            result['DiskType'] = self.disk_type
+        if self.eip_max is not None:
+            result['EipMax'] = self.eip_max
+        if self.io_max is not None:
+            result['IoMax'] = self.io_max
+        if self.io_max_spec is not None:
+            result['IoMaxSpec'] = self.io_max_spec
+        if self.paid_type is not None:
+            result['PaidType'] = self.paid_type
+        if self.partition_num is not None:
+            result['PartitionNum'] = self.partition_num
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+        if self.resource_group_id is not None:
+            result['ResourceGroupId'] = self.resource_group_id
+        if self.serverless_config is not None:
+            result['ServerlessConfig'] = self.serverless_config.to_map()
+        if self.spec_type is not None:
+            result['SpecType'] = self.spec_type
+        result['Tag'] = []
+        if self.tag is not None:
+            for k in self.tag:
+                result['Tag'].append(k.to_map() if k else None)
+        if self.topic_quota is not None:
+            result['TopicQuota'] = self.topic_quota
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DeployType') is not None:
+            self.deploy_type = m.get('DeployType')
+        if m.get('DiskSize') is not None:
+            self.disk_size = m.get('DiskSize')
+        if m.get('DiskType') is not None:
+            self.disk_type = m.get('DiskType')
+        if m.get('EipMax') is not None:
+            self.eip_max = m.get('EipMax')
+        if m.get('IoMax') is not None:
+            self.io_max = m.get('IoMax')
+        if m.get('IoMaxSpec') is not None:
+            self.io_max_spec = m.get('IoMaxSpec')
+        if m.get('PaidType') is not None:
+            self.paid_type = m.get('PaidType')
+        if m.get('PartitionNum') is not None:
+            self.partition_num = m.get('PartitionNum')
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+        if m.get('ResourceGroupId') is not None:
+            self.resource_group_id = m.get('ResourceGroupId')
+        if m.get('ServerlessConfig') is not None:
+            temp_model = CreatePostPayOrderRequestServerlessConfig()
+            self.serverless_config = temp_model.from_map(m['ServerlessConfig'])
+        if m.get('SpecType') is not None:
+            self.spec_type = m.get('SpecType')
+        self.tag = []
+        if m.get('Tag') is not None:
+            for k in m.get('Tag'):
+                temp_model = CreatePostPayOrderRequestTag()
+                self.tag.append(temp_model.from_map(k))
+        if m.get('TopicQuota') is not None:
+            self.topic_quota = m.get('TopicQuota')
+        return self
+
+
+class CreatePostPayOrderShrinkRequestTag(TeaModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+    ):
+        # The key of tag N.
+        # 
+        # *   Valid values of N: 1 to 20.
+        # *   If this parameter is left empty, the keys of all tags are matched.
+        # *   The tag key must be up to 128 characters in length. It cannot start with acs: or aliyun or contain [http:// or https://.](http://https://。)
+        self.key = key
+        # The value of tag N.
+        # 
+        # *   Valid values of N: 1 to 20.
+        # *   If you do not specify a tag key, you cannot specify a tag value. If this parameter is not configured, all tag values are matched.
+        # *   The tag value must be 1 to 128 characters in length. It cannot start with acs: or aliyun or contain [http:// or https://.](http://https://。)
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.key is not None:
+            result['Key'] = self.key
+        if self.value is not None:
+            result['Value'] = self.value
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+        return self
+
+
+class CreatePostPayOrderShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        deploy_type: int = None,
+        disk_size: int = None,
+        disk_type: str = None,
+        eip_max: int = None,
+        io_max: int = None,
+        io_max_spec: str = None,
+        paid_type: int = None,
+        partition_num: int = None,
+        region_id: str = None,
+        resource_group_id: str = None,
+        serverless_config_shrink: str = None,
+        spec_type: str = None,
+        tag: List[CreatePostPayOrderShrinkRequestTag] = None,
+        topic_quota: int = None,
+    ):
+        # The deployment mode of the instance. Valid values:
+        # 
+        # *   **4**: deploys the instance that allows access from the Internet and a VPC.
+        # *   **5**: deploys the instance that allows access only from a VPC.
+        self.deploy_type = deploy_type
+        # The disk size.
+        # 
+        # For more information about the valid values, see [Billing](~~84737~~).
+        self.disk_size = disk_size
+        # The disk type. Valid values:
+        # 
+        # *   **0**: ultra disk
+        # *   **1**: standard SSD
+        self.disk_type = disk_type
+        # The Internet traffic for the instance.
+        # 
+        # *   This parameter is required if the **DeployType** parameter is set to **4**.
+        # *   For more information about the valid values, see [Billing](~~84737~~).
+        self.eip_max = eip_max
+        # The maximum traffic for the instance. We recommend that you do not configure this parameter.
+        # 
+        # *   You must specify at least one of the IoMax and IoMaxSpec parameters. If you configure both parameters, the value of the IoMaxSpec parameter takes effect. We recommend that you specify only the IoMaxSpec parameter.
+        # *   For more information about the valid values, see [Billing](~~84737~~).
+        self.io_max = io_max
+        # The traffic specification of the instance. We recommend that you configure this parameter.
+        # 
+        # *   You must specify at least one of the IoMax and IoMaxSpec parameters. If you configure both parameters, the value of the IoMaxSpec parameter takes effect. We recommend that you specify only the IoMaxSpec parameter.
+        # *   For more information about the valid values, see [Billing](~~84737~~).
+        self.io_max_spec = io_max_spec
+        self.paid_type = paid_type
+        # The number of partitions. We recommend that you configure this parameter.
+        # 
+        # *   You must specify at least one of the PartitionNum and TopicQuota parameters. We recommend that you configure only the PartitionNum parameter.
+        # *   If you specify both parameters, the topic-based sales model is used to check whether the PartitionNum value and the TopicQuota value are the same. If they are not the same, a failure response is returned. If they are the same, the order is placed based on the PartitionNum value.
+        # *   For more information about the valid values, see [Billing](~~84737~~).
+        self.partition_num = partition_num
+        # The region ID of the instance.
+        self.region_id = region_id
+        # The ID of the resource group.
+        # 
+        # If this parameter is left empty, the default resource group is used. You can view the resource group ID on the Resource Group page in the Resource Management console.
+        self.resource_group_id = resource_group_id
+        self.serverless_config_shrink = serverless_config_shrink
         # The edition of the instance. Valid values:
         # 
         # *   **normal**: Standard Edition (High Write)
@@ -827,12 +1069,16 @@ class CreatePostPayOrderRequest(TeaModel):
             result['IoMax'] = self.io_max
         if self.io_max_spec is not None:
             result['IoMaxSpec'] = self.io_max_spec
+        if self.paid_type is not None:
+            result['PaidType'] = self.paid_type
         if self.partition_num is not None:
             result['PartitionNum'] = self.partition_num
         if self.region_id is not None:
             result['RegionId'] = self.region_id
         if self.resource_group_id is not None:
             result['ResourceGroupId'] = self.resource_group_id
+        if self.serverless_config_shrink is not None:
+            result['ServerlessConfig'] = self.serverless_config_shrink
         if self.spec_type is not None:
             result['SpecType'] = self.spec_type
         result['Tag'] = []
@@ -857,18 +1103,22 @@ class CreatePostPayOrderRequest(TeaModel):
             self.io_max = m.get('IoMax')
         if m.get('IoMaxSpec') is not None:
             self.io_max_spec = m.get('IoMaxSpec')
+        if m.get('PaidType') is not None:
+            self.paid_type = m.get('PaidType')
         if m.get('PartitionNum') is not None:
             self.partition_num = m.get('PartitionNum')
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
         if m.get('ResourceGroupId') is not None:
             self.resource_group_id = m.get('ResourceGroupId')
+        if m.get('ServerlessConfig') is not None:
+            self.serverless_config_shrink = m.get('ServerlessConfig')
         if m.get('SpecType') is not None:
             self.spec_type = m.get('SpecType')
         self.tag = []
         if m.get('Tag') is not None:
             for k in m.get('Tag'):
-                temp_model = CreatePostPayOrderRequestTag()
+                temp_model = CreatePostPayOrderShrinkRequestTag()
                 self.tag.append(temp_model.from_map(k))
         if m.get('TopicQuota') is not None:
             self.topic_quota = m.get('TopicQuota')
@@ -5137,7 +5387,9 @@ class GetInstanceListResponseBodyInstanceListInstanceVO(TeaModel):
         expired_time: int = None,
         instance_id: str = None,
         io_max: int = None,
+        io_max_read: int = None,
         io_max_spec: str = None,
+        io_max_write: int = None,
         kms_key_id: str = None,
         msg_retain: int = None,
         name: str = None,
@@ -5199,8 +5451,10 @@ class GetInstanceListResponseBodyInstanceListInstanceVO(TeaModel):
         self.instance_id = instance_id
         # The peak traffic allowed for the instance.
         self.io_max = io_max
+        self.io_max_read = io_max_read
         # The traffic specification.
         self.io_max_spec = io_max_spec
+        self.io_max_write = io_max_write
         # The ID of the key that is used for disk encryption in the region where the instance is deployed.
         self.kms_key_id = kms_key_id
         # The retention period of messages in the instance. Unit: hours.
@@ -5319,8 +5573,12 @@ class GetInstanceListResponseBodyInstanceListInstanceVO(TeaModel):
             result['InstanceId'] = self.instance_id
         if self.io_max is not None:
             result['IoMax'] = self.io_max
+        if self.io_max_read is not None:
+            result['IoMaxRead'] = self.io_max_read
         if self.io_max_spec is not None:
             result['IoMaxSpec'] = self.io_max_spec
+        if self.io_max_write is not None:
+            result['IoMaxWrite'] = self.io_max_write
         if self.kms_key_id is not None:
             result['KmsKeyId'] = self.kms_key_id
         if self.msg_retain is not None:
@@ -5400,8 +5658,12 @@ class GetInstanceListResponseBodyInstanceListInstanceVO(TeaModel):
             self.instance_id = m.get('InstanceId')
         if m.get('IoMax') is not None:
             self.io_max = m.get('IoMax')
+        if m.get('IoMaxRead') is not None:
+            self.io_max_read = m.get('IoMaxRead')
         if m.get('IoMaxSpec') is not None:
             self.io_max_spec = m.get('IoMaxSpec')
+        if m.get('IoMaxWrite') is not None:
+            self.io_max_write = m.get('IoMaxWrite')
         if m.get('KmsKeyId') is not None:
             self.kms_key_id = m.get('KmsKeyId')
         if m.get('MsgRetain') is not None: